Started on testing function + added samples
Samples are holding the indices of the current state of the data. They are initialized to be 0:num_data and when the data is shuffled for multi-class they get shuffled too, to preserve the state. This will later be used for Numeric class which will do the predictions at the end.